Solved shared managers
Able to share the vole class between processes (need to figure out to create proxies that access the right referent). Now I need to globalize all the variables in the main() function of both the rfid_main and doors_main so that I can have multiprocess spawn them as child processes.